// src/modules/auto-increment/auto-increment.enum.ts
export enum AutoIncrementEnum {
    USER = 'U',
    ENTITY = 'EN',
    STUDENT = 'ST',
    COURSE = 'CO',
    EMPLOYEE = 'EM',
    ROLE = 'RO',
    PERMISSION = 'PE',
    SUBJECT = 'SU',
    ASSIGNMENT = 'AS',
    SOLUTION = 'ASO',
    ANNOUNCEMENT = 'AN',
    SUBSCRIPTION = 'SUB',
    ENROLLMENT = 'ENROL'
}