// src/enums/module-type.enum.ts
export enum ModuleType {
    User = "User",
    Company = "Company",
    College = "College",
    Course = "Course",
    Employee = "Employee",
    Student = "Student",
    ProblemStatement = "Problem Statement",
    Announcement = "Announcement",
    Subscription = "Subscription",
    Role = "Role"
}